+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#222] Support max_age=0
 - [#221] Avoid raising invalid_request error on prompt=create
 - [#220] Define priority on possible prompt values to statically & successfully process multiple prompt values
+- [#224] Define priority between max_age & prompt
 
 ## v1.8.10 (2024-11-29)